feat: DRY error list components and styles across devices
